Prevent atext atoms from backtracking; atoms should be fully greedy
This prevents a quadratic exposion in the 'obs-phrase' regular
expression.  In reduced form, 'obs-phrase' is equivalent to 'atom*',
which is '(atext+)*'.  In the context of 'obs-mailbox', this comes
before a required '<' -- but due to the alternation with obs-addr-spec,
the optimizer cannot short-circuit the match due to a lack of '<'
characters in the string.  As such, it tries every possible different
possibility of splitting the phrase into one-or-more character atoms.
This leads to a quadratic growth in the length of the string.

As the 'atext' class is intentionally constructed to never contain any
"interesting" characters upon which alternation might hinge,
backtracking through them is never fruitful -- force the match on atoms
to be non-backtracking.